requirements.txt
## top-level packages
arrow==1.3.0
ics==0.7.2
python-dotenv==1.0.0

## dependencies:
attrs==23.1.0
python-dateutil==2.8.2
six==1.16.0
TatSu==5.11.2
types-python-dateutil==2.8.19.14
